Change PacketBuffer size to int as it should have been in the first place.
git-svn-id: http://svn.mythtv.org/svn/trunk@26540 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythtv/privatedecoder_crystalhd.cpp
@@ -415,7 +415,7 @@ int PrivateDecoderCrystalHD::ProcessPacket(AVStream *stream, AVPacket *pkt)
     {
 
         PacketBuffer *buffer = m_packet_buffers.last();
-        if (GetTxFreeSize(0) < (int)buffer->size)
+        if (GetTxFreeSize(0) < buffer->size)
         {
             usleep(10000);
             return 0;
